Move win32 specific file to the right dir
