+++ src/win32/edge_utils_win32.h
